delete old reingold version
