+++ raphtory-graphql/src/model/algorithms/mod.rs
@@ -5,7 +5,6 @@ pub mod algorithm;
 pub mod algorithm_entry_point;
 pub mod document;
 pub mod graph_algorithms;
-pub mod layout_algorithm;
 pub mod similarity_search;
 pub mod vector_algorithms;
