+++ raphtory/src/python/packages/algorithms.rs
@@ -15,7 +15,7 @@ use crate::{
         dynamics::temporal::epidemics::{temporal_SEIR as temporal_SEIR_rs, Infected, SeedError},
         layout::{
             cohesive_fruchterman_reingold::cohesive_fruchterman_reingold as cohesive_fruchterman_reingold_rs,
-            fruchterman_reingold_unbounded::fruchterman_reingold_unbounded as fruchterman_reingold_rs,
+            fruchterman_reingold::fruchterman_reingold_unbounded as fruchterman_reingold_rs,
         },
         metrics::{
             balance::balance as balance_rs,
@@ -691,31 +691,47 @@ pub fn louvain(
 }
 
 #[pyfunction]
-#[pyo3[signature=(graph, iterations=100, scale=1.0, cooloff_factor=0.95, dt=0.1)]]
+#[pyo3[signature=(graph, iterations=100, scale=1.0, node_start_size=1.0, cooloff_factor=0.95, dt=0.1)]]
 pub fn fruchterman_reingold(
     graph: &PyGraphView,
     iterations: u64,
     scale: f32,
+    node_start_size: f32,
     cooloff_factor: f32,
     dt: f32,
 ) -> HashMap<u64, [f32; 2]> {
-    fruchterman_reingold_rs(&graph.graph, iterations, scale, cooloff_factor, dt)
-        .into_iter()
-        .map(|(id, vector)| (id, [vector.x, vector.y]))
-        .collect()
+    fruchterman_reingold_rs(
+        &graph.graph,
+        iterations,
+        scale,
+        node_start_size,
+        cooloff_factor,
+        dt,
+    )
+    .into_iter()
+    .map(|(id, vector)| (id, [vector.x, vector.y]))
+    .collect()
 }
 
 #[pyfunction]
-#[pyo3[signature=(graph, iterations=100, scale=1.0, cooloff_factor=0.95, dt=0.1)]]
+#[pyo3[signature=(graph, iterations=100, scale=1.0, node_start_size=1.0, cooloff_factor=0.95, dt=0.1)]]
 pub fn cohesive_fruchterman_reingold(
     graph: &PyGraphView,
     iterations: u64,
     scale: f32,
+    node_start_size: f32,
     cooloff_factor: f32,
     dt: f32,
 ) -> HashMap<u64, [f32; 2]> {
-    cohesive_fruchterman_reingold_rs(&graph.graph, iterations, scale, cooloff_factor, dt)
-        .into_iter()
-        .map(|(id, vector)| (id, [vector.x, vector.y]))
-        .collect()
+    cohesive_fruchterman_reingold_rs(
+        &graph.graph,
+        iterations,
+        scale,
+        node_start_size,
+        cooloff_factor,
+        dt,
+    )
+    .into_iter()
+    .map(|(id, vector)| (id, [vector.x, vector.y]))
+    .collect()
 }
